+++ CHANGELOG.md
@@ -6,6 +6,23 @@ All user visible changes to this project will be documented in this file. This p
 
 
 
+## [0.1.0-alpha.14] · 2024-??-??
+[0.1.0-alpha.14]: /../../tree/v0.1.0-alpha.14
+
+[Diff](/../../compare/v0.1.0-alpha.13.2...v0.1.0-alpha.14) | [Milestone](/../../milestone/21)
+
+### Added
+
+- UI:
+    - Account deletion page. ([#944])
+    - Terms and conditions. ([#944])
+    - Privacy policy. ([#944])
+
+[#944]: /../../pull/944
+
+
+
+
 ## [0.1.0-alpha.13.2] · 2024-04-11
 [0.1.0-alpha.13.2]: /../../tree/v0.1.0-alpha.13.2